pyproject.toml
[project]
name = "django-rest-framework-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "groq>=0.32.0",
    "psycopg2>=2.9.11",
]
